refactor runner to command
